refactor: Styleguide files
